# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: PrintFlow 3D API
  description: |
    API REST de PrintFlow — Gestion d'impressions 3D et filaments.

    ## Authentification
    Si l'authentification est activée dans les paramètres, ajoutez le header :
    ```
    X-Auth-Token: votre_mot_de_passe
    ```

    ## Intégration Home Assistant
    Exemple de sensor REST pour le stock de filaments :
    ```yaml
    sensor:
      - platform: rest
        name: PrintFlow stock faible
        resource: http://192.168.1.145/api/stats/alerts
        method: GET
        value_template: "{{ value_json | length }}"
        unit_of_measurement: "bobines"
    ```

  version: "1.7.0"
  contact:
    name: PrintFlow 3D
    url: https://github.com/Fabdub59650/printflow-3d
servers:
  - url: /api
    description: Serveur local (Raspberry Pi)
...
paths:
...
  /filaments:
    get:
      tags: [Filaments]
      summary: Lister les filaments
      parameters:
        - name: archived
          in: query
          description: "1 pour inclure les bobines archivées"
          schema:
            type: integer
            enum: [0, 1]
      responses:
        "200":
          description: Liste des filaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Filament'
              example:
                - id: 1
                  name: "PLA Blanc Elegoo"
                  brand: "Elegoo"
                  material: "PLA"
                  color_hex: "#FFFFFF"
                  color_name: "Blanc"
                  weight_remaining: 850
                  weight_total: 1000
                  temp_nozzle_min: 190
                  temp_nozzle_max: 220
                  archived: 0
    post:
      tags: [Filaments]
      summary: Créer un filament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilamentInput'
            example:
              name: "PETG Noir"
              brand: "Bambu"
              material: "PETG"
              color_hex: "#000000"
              color_name: "Noir"
              weight_total: 1000
              weight_remaining: 1000
              temp_nozzle_min: 230
              temp_nozzle_max: 250
      responses:
        "201":
          description: Filament créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Filament'
...
  /prints:
    get:
      tags: [Impressions]
      summary: Lister les impressions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [printing, done, failed, paused, queued, cancelled]
        - name: printer_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Liste des impressions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Print'
              example:
                - id: 1
                  name: "Support caméra"
                  status: "done"
                  filament_used: 45
                  actual_duration: 120
                  printer_name: "Elegoo Centauri"
                  filament_name: "PLA Blanc"
                  color_hex: "#FFFFFF"
    post:
      tags: [Impressions]
      summary: Créer une impression
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrintInput'
            example:
              name: "Pied de lampe"
              printer_id: 1
              filament_id: 2
              status: "printing"
              estimated_duration: 180
      responses:
        "201":
          description: Impression créée
...
  /printers:
    get:
      tags: [Imprimantes]
      summary: Lister les imprimantes
      responses:
        "200":
          description: Liste des imprimantes
          content:
            application/json:
              example:
                - id: 1
                  name: "Elegoo Centauri Carbon 2"
                  model: "Centauri Carbon 2"
                  status: "idle"
                  interface_url: "http://192.168.1.200"
    post:
      tags: [Imprimantes]
      summary: Ajouter une imprimante
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Bambu X1C"
              model: "X1 Carbon"
              status: "idle"
      responses:
        "201":
          description: Imprimante créée
...
  /maintenance:
    get:
      tags: [Maintenance]
      summary: Lister les entrées de maintenance
      responses:
        "200":
          description: Liste des maintenances
          content:
            application/json:
              example:
                - id: 1
                  printer_name: "Elegoo Centauri"
                  type: "nettoyage"
                  performed_at: "2026-04-01"
                  next_due: "2026-05-01"
                  description: "Nettoyage buse et plateau"
    post:
      tags: [Maintenance]
      summary: Créer une entrée de maintenance
      requestBody:
        required: true
        content:
          application/json:
            example:
              printer_id: 1
              type: "nettoyage"
              performed_at: "2026-04-08"
              next_due: "2026-05-08"
              description: "Nettoyage complet"
      responses:
        "201":
          description: Maintenance créée
...
  /stats/consumption:
    get:
      tags: [Statistiques]
      summary: Consommation de filament
      parameters:
        - name: days
          in: query
          description: Nombre de jours en arrière (7, 30 ou 90)
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Données de consommation
          content:
            application/json:
              example:
                byMaterial:
                  - material: "PLA"
                    total_g: 1250
                  - material: "PETG"
                    total_g: 380
                byDay:
                  - date: "2026-04-08"
                    total_g: 145
...
  /history:
    get:
      tags: [Historique]
      summary: Journal des modifications
      parameters:
        - name: entity
          in: query
          description: Filtrer par type d'entité
          schema:
            type: string
            enum: [filament, print, printer, project, library_object, nfc]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Entrées du journal
          content:
            application/json:
              example:
                - id: 1
                  entity: "filament"
                  entity_id: 3
                  action: "create"
                  detail: "Filament créé : PLA Rouge"
                  created_at: "2026-04-08T10:30:00Z"

  # ── NFC ───────────────────────────────────────────────────
...
  /settings:
    get:
      tags: [Paramètres]
      summary: Lire les paramètres
      responses:
        "200":
          description: Paramètres de l'application
          content:
            application/json:
              example:
                app_name: "PrintFlow"
                theme: "blue"
                stock_alert_enabled: "true"
                stock_alert_threshold: "20"
                maintenance_alert_enabled: "true"
                maintenance_alert_days: "30"
    put:
      tags: [Paramètres]
      summary: Modifier les paramètres
      requestBody:
        required: true
        content:
          application/json:
            example:
              stock_alert_threshold: "15"
              maintenance_alert_days: "14"
      responses:
        "200":
          description: Paramètres mis à jour
components:
...
  schemas:
    Filament:
      type: object
      properties:
        id:          { type: integer }
        name:        { type: string }
        brand:       { type: string }
        material:    { type: string, enum: [PLA, PETG, ABS, ASA, TPU, Nylon, PC, HIPS, PVA, autre] }
        color_hex:   { type: string, example: "#3b82f6" }
        color_name:  { type: string }
        diameter:    { type: number, example: 1.75 }
        weight_total:     { type: number }
        weight_remaining: { type: number }
        spool_weight:     { type: number }
        temp_nozzle_min:  { type: integer }
        temp_nozzle_max:  { type: integer }
        temp_bed_min:     { type: integer }
        temp_bed_max:     { type: integer }
        price:       { type: number }
        location:    { type: string }
        notes:       { type: string }
        nfc_uid:     { type: string }
        archived:    { type: integer, enum: [0, 1] }
    FilamentInput:
      type: object
      required: [name, material]
      properties:
        name:        { type: string }
        brand:       { type: string }
        material:    { type: string }
        color_hex:   { type: string }
        color_name:  { type: string }
        diameter:    { type: number }
        weight_total:     { type: number }
        weight_remaining: { type: number }
        temp_nozzle_min:  { type: integer }
        temp_nozzle_max:  { type: integer }
        temp_bed_min:     { type: integer }
        temp_bed_max:     { type: integer }
        price:       { type: number }
        location:    { type: string }
        notes:       { type: string }
    Print:
      type: object
      properties:
        id:               { type: integer }
        name:             { type: string }
        status:           { type: string, enum: [printing, done, failed, paused, queued, cancelled] }
        printer_id:       { type: integer }
        filament_id:      { type: integer }
        filament_used:    { type: number, description: "Grammes consommés" }
        estimated_duration: { type: integer, description: "Minutes" }
        actual_duration:    { type: integer, description: "Minutes" }
        progress:         { type: integer, description: "Pourcentage 0-100" }
        file_name:        { type: string }
        notes:            { type: string }
        created_at:       { type: string, format: date-time }
    PrintInput:
      type: object
      required: [name]
      properties:
        name:             { type: string }
        printer_id:       { type: integer }
        filament_id:      { type: integer }
        status:           { type: string }
        estimated_duration: { type: integer }
        actual_duration:    { type: integer }
        filament_used:    { type: number }
        progress:         { type: integer }
        file_name:        { type: string }
        notes:            { type: string }
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: Requis uniquement si l'authentification est activée dans les paramètres
security:
  - ApiKeyAuth: []
  - {}